Fix float error

+++ 0x10-variadic_functions/3-print_all.c
@@ -39,7 +39,7 @@ void print_float(va_list arg)
 {
 	float n;
 
-	n = va_arg(arg, int);
+	n = va_arg(arg, double);
 	printf("%f", n);
 }
